# api/openapi/swagger.yaml
basePath: /
definitions:
...
  DataResponse:
    properties:
      data: {}
    type: object
...
  ListResponse:
    properties:
      cursor:
        type: string
      data: {}
    type: object
...
  MatchDetailResponse:
    properties:
      ceremony_type:
        type: string
      event_address:
        type: string
      event_date:
        type: string
      event_id:
        type: string
      id:
        type: string
      matched_at:
        type: string
      pandit_first_name:
        type: string
      pandit_id:
        type: string
      pandit_last_name:
        type: string
      pandit_phone:
        type: string
      pandit_photo_url:
        type: string
      status:
        type: string
      yajman_first_name:
        type: string
      yajman_id:
        type: string
      yajman_last_name:
        type: string
      yajman_phone:
        type: string
      yajman_photo_url:
        type: string
    type: object
...
  ProblemDetail:
    properties:
      detail:
        type: string
      instance:
        type: string
      status:
        type: integer
      title:
        type: string
      type:
        type: string
    type: object
...
info:
  contact:
    email: support@medha.app
    name: Medha Backend Team
    url: https://medha.app
  description: Backend API for the Medha platform. V2 Authentication (OTP-based),
    User Profiles, Events, Matching, and more.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://swagger.io/terms/
  title: Medha API V2
  version: 2.0.0
paths:
...
  /api/v2/leads/{id}/accept:
    put:
      description: Pandit accepts a job lead. Platform fee is charged.
      parameters:
      - description: Lead ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lead accepted
          schema:
            $ref: '#/definitions/DataResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/ProblemDetail'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ProblemDetail'
      security:
      - BearerAuth: []
      summary: Accept a job lead
      tags:
      - lead
  /api/v2/leads/{id}/decline:
    put:
      description: Pandit declines a job lead. No fee is charged.
      parameters:
      - description: Lead ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lead declined
          schema:
            $ref: '#/definitions/DataResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/ProblemDetail'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ProblemDetail'
      security:
      - BearerAuth: []
      summary: Decline a job lead
      tags:
      - lead
  /api/v2/leads/mine:
    get:
      description: Pandit views their incoming job leads.
      produces:
      - application/json
      responses:
        "200":
          description: Leads list
          schema:
            $ref: '#/definitions/DataResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ProblemDetail'
      security:
      - BearerAuth: []
      summary: List my leads
      tags:
      - lead
...
  /api/v2/match/{id}:
    get:
      description: Returns the full details of a match including both parties' info.
        Only the yajman or pandit can access.
      parameters:
      - description: Match UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Match details
          schema:
            allOf:
            - $ref: '#/definitions/DataResponse'
            - properties:
                data:
                  $ref: '#/definitions/MatchDetailResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ProblemDetail'
        "403":
          description: Forbidden — not a party
          schema:
            $ref: '#/definitions/ProblemDetail'
        "404":
          description: Match not found
          schema:
            $ref: '#/definitions/ProblemDetail'
      security:
      - BearerAuth: []
      summary: Get match details
      tags:
      - match-v2
  /api/v2/match/{id}/accept:
    put:
      description: Transitions a match from 'created' to 'matched'. Only the yajman
        can accept.
      parameters:
      - description: Match UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Match accepted
          schema:
            allOf:
            - $ref: '#/definitions/DataResponse'
            - properties:
                data: {}
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ProblemDetail'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ProblemDetail'
        "404":
          description: Match not found
          schema:
            $ref: '#/definitions/ProblemDetail'
      security:
      - BearerAuth: []
      summary: Accept match
      tags:
      - match-v2
  /api/v2/match/{id}/activate:
    put:
      description: Transitions a match from 'matched' to 'active' (ceremony in progress).
      parameters:
      - description: Match UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Match activated
          schema:
            allOf:
            - $ref: '#/definitions/DataResponse'
            - properties:
                data: {}
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ProblemDetail'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ProblemDetail'
        "404":
          description: Match not found
          schema:
            $ref: '#/definitions/ProblemDetail'
      security:
      - BearerAuth: []
      summary: Activate match
      tags:
      - match-v2
  /api/v2/match/{id}/complete:
    put:
      description: Transitions a match from 'active' to 'completed'. Only the yajman
        can complete.
      parameters:
      - description: Match UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Match completed
          schema:
            allOf:
            - $ref: '#/definitions/DataResponse'
            - properties:
                data: {}
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ProblemDetail'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/ProblemDetail'
        "404":
          description: Match not found
          schema:
            $ref: '#/definitions/ProblemDetail'
      security:
      - BearerAuth: []
      summary: Complete match
      tags:
      - match-v2
...
  /api/v2/match/mine:
    get:
      description: Returns paginated matches for the authenticated user (as yajman
        or pandit).
      parameters:
      - description: Pagination cursor
        in: query
        name: cursor
        type: string
      - description: Page size (default 20)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: My matches
          schema:
            allOf:
            - $ref: '#/definitions/ListResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/MatchDetailResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ProblemDetail'
      security:
      - BearerAuth: []
      summary: List my matches
      tags:
      - match-v2
...
securityDefinitions:
  AdminToken:
    description: Type "Bearer" followed by a space and an admin JWT or break-glass
      token.
    in: header
    name: Authorization
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...